refactor: throttle 함수를 useThrottle함수로 개선 - 리랜더링시 함수가 초기화되면서 throttle함수의 timer와 callTime이 제대로 관리되지 못하는 문제를 해결했다
